feat: restructure module imports to enhance path accessibility and improve code organization
