+++ scripts/cisco/12.2.2/modules/fabric/delete_fabric.py
@@ -14,7 +14,7 @@ from pathlib import Path
 # Add parent directory to path to access api and config_utils
 sys.path.append(str(Path(__file__).parent.parent.absolute()))
 import api.fabric as fabric_api
-from config_utils import print_build_summary
+from modules.config_utils import print_build_summary
 
 class FabricDeleter:
     """Handles fabric deletion operations."""
